[gig] client content update

+++ web/_gig/dev/clients/OzMinerals/OzMinerals.php
@@ -27,7 +27,7 @@ class OzMinerals extends Client
         return new DateTimeZone('Australia/Perth');
     }
 
-    public function ShiftStart(): int
+    public function ShiftStart(): float
     {
         return 7;
     }
